requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
beniget==0.4.1
bleach==6.1.0
blinker==1.8.2
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.0.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
chardet==4.0.0
charset-normalizer==3.4.0
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0
django-cors-headers==4.4.0
django-sslserver==0.22
duplicity==0.8.21
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
fasteners==0.14.1
fastjsonschema==2.20.0
Flask==2.3.2
fonttools==4.53.1
fqdn==1.5.1
fs==2.4.12
future==0.18.2
gast==0.5.2
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==1.7.11
google-auth==2.36.0
google-auth-httplib2==0.1.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==1.1.2
grpcio==1.67.1
grpcio-status==1.67.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.28.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-console==6.4.0
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
mysqlclient==2.1.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.0.1
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==4.3.6
ply==3.11
powerline-status==2.7
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.9.2
pydantic_core==2.23.4
PyDrive==1.3.1
pygame==2.6.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-json-logger==2.0.7
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.2.0
razorpay==1.4.2
referencing==0.35.1
reportlab==3.6.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
rsa==4.8
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
selenium==4.24.0
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.31
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.9
systemd-python==234
terminado==0.18.1
testpath==0.5.0
tinycss2==1.3.0
tomli==2.0.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.3.0
uritemplate==3.0.1
urllib3==1.26.5
urwid==2.1.2
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==2.0.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
